agent_suite_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
chromadb
sentence-transformers
pandas